Edited
Simon,
This is a much more complicated thing to do and the suggested links to categories in the existing table cannot be done in Markdown. I created a tabbed table that just 2 weeks ago became available in Markdown for IBM Cloud Docs. I tested it in staging and will be promoting that version up to production. Essentially, I'm discarding this PR for now and will immediately promote the version that I tested in staging.